requirements.txt
aiohttp==3.7.4.post0
asgiref==3.3.4
async-timeout==3.0.1
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
dateparser==1.0.0
Django==3.2.4
django-cors-headers==3.7.0
django-environ==0.4.5
djangorestframework==3.12.4
idna==2.10
multidict==5.1.0
psycopg2==2.9.1
python-binance==1.0.12
python-dateutil==2.8.1
pytz==2021.1
regex==2021.4.4
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
tzlocal==2.1
ujson==4.0.2
urllib3==1.26.6
websockets==9.1
yarl==1.6.3
